refactor przeniesienie funkcji
