+++ src/app/energy-summary-list/energy-summary-list.component.html
@@ -85,7 +85,7 @@
 </table>
 
 <h3>Uwagi</h3>
-<p>Net mettering posiada dene od stycznia 2023 do wrzesnia 2024</p>
+<p>Net mettering posiada dene od czerwca 2022 do wrzesnia 2024</p>
 <p>
   Ceny obejmuja tylko dystrybucje i cene za energie elektryczna, bez kosztow
   typy oplata mocowa, oplata oze itd. Sa szacunkowe.
