+++ src/app/energy-summary-list/prosumer-type-selector/prosumer-type-selector.component.html
@@ -13,7 +13,7 @@
       Net Billing godzinowy
     </mat-radio-button> -->
     <mat-radio-button [value]="prosumerTypes.netBillingMonthly">
-      Net Billing miesięczny
+      Net Billing miesięczn
     </mat-radio-button>
   </mat-radio-group>
 </form>
