+++ src/app/store/energy-data-store/prosument-type.enum.ts
@@ -1,6 +1,6 @@
 export enum ProsumerType {
-    netBillingMonthly = 'netBillingMonthly',
-    netBillingHourly  = 'netBillingHourly',
-    netMetering80 = 'netMetering80',
-    netMetering70= 'netMetering70',
+  netBillingMonthly = 'netBillingMonthly',
+  netBillingHourly = 'netBillingHourly',
+  netMetering80 = 'netMetering80',
+  netMetering70 = 'netMetering70',
 }
